changed ReadHrs and ReadAls to uint32, and did static_cast instead of hidden cast when using it

+++ src/drivers/Hrs3300.cpp
@@ -54,14 +54,14 @@ void Hrs3300::Disable() {
   WriteRegister(static_cast<uint8_t>(Registers::Enable), value);
 }
 
-uint16_t Hrs3300::ReadHrs() {
+uint32_t Hrs3300::ReadHrs() {
   auto m = ReadRegister(static_cast<uint8_t>(Registers::C0DataM));
   auto h = ReadRegister(static_cast<uint8_t>(Registers::C0DataH));
   auto l = ReadRegister(static_cast<uint8_t>(Registers::C0dataL));
   return ((l & 0x30) << 12) | (m << 8) | ((h & 0x0f) << 4) | (l & 0x0f);
 }
 
-uint16_t Hrs3300::ReadAls() {
+uint32_t Hrs3300::ReadAls() {
   auto m = ReadRegister(static_cast<uint8_t>(Registers::C1dataM));
   auto h = ReadRegister(static_cast<uint8_t>(Registers::C1dataH));
   auto l = ReadRegister(static_cast<uint8_t>(Registers::C1dataL));
